Ajustes na Tela de Login, e Cadastro conforme especicações no MarvelApp, extração de valores e textos para os arquivos de dimens e strings. Inclusão de resources de imagens e logos. Mudança dos nomes das variaveis e classes para inglês.
