--- app/src/main/java/com/rafaelmfer/nasaexperience/CadastroActivity.java
@@ -1,29 +1,0 @@
-package com.rafaelmfer.nasaexperience;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-import static com.rafaelmfer.nasaexperience.R.id.*;
-
-public class CadastroActivity extends AppCompatActivity {
-
-    Button btnVoltar;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_cadastro);
-        btnVoltar = findViewById(R.id.btnVoltar);
-
-        btnVoltar.setOnClickListener(new View.OnClickListener(){
-           @Override
-            public void onClick(View v){
-               onBackPressed();
-           }
-
-        });
-    }
-}
