+++ app/src/main/java/com/rafaelmfer/nasaexperience/LoginActivity.java
@@ -1,0 +1,33 @@
+package com.rafaelmfer.nasaexperience;
+
+import android.content.Intent;
+import android.os.Bundle;
+import android.view.View;
+import android.widget.Button;
+
+import androidx.appcompat.app.AppCompatActivity;
+
+import com.google.android.material.textfield.TextInputLayout;
+
+public class LoginActivity extends AppCompatActivity {
+
+    TextInputLayout tilLoginUserName, tilLoginUserPassword;
+    Button btSignIn, btFacebookSignIn, btGoogleSignIn, btSignUp;
+
+    @Override
+    protected void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+        setContentView(R.layout.activity_login);
+
+        btSignUp = findViewById(R.id.btSignUp);
+
+        btSignUp.setOnClickListener(new View.OnClickListener() {
+            @Override
+            public void onClick(View view) {
+                Intent intent = new Intent(view.getContext(), RegisterActivity.class);
+                startActivity(intent);
+            }
+        });
+
+    }
+}
