--- app/src/main/java/com/rafaelmfer/nasaexperience/MainActivity.java
@@ -1,37 +1,0 @@
-package com.rafaelmfer.nasaexperience;
-
-import androidx.appcompat.app.AppCompatActivity;
-
-import android.content.Intent;
-import android.os.Bundle;
-import android.view.View;
-import android.widget.Button;
-
-import com.google.android.material.textfield.TextInputLayout;
-
-public class MainActivity extends AppCompatActivity {
-   TextInputLayout usuario;
-   TextInputLayout senha;
-   Button btnEntrar;
-   Button btnFacebook;
-   Button btnGoogle;
-   Button btnCadastrar;
-
-    @Override
-    protected void onCreate(Bundle savedInstanceState) {
-        super.onCreate(savedInstanceState);
-        setContentView(R.layout.activity_main);
-
-        btnCadastrar = (Button) findViewById(R.id.btnCadastrar);
-
-        btnCadastrar.setOnClickListener(new View.OnClickListener() {
-            @Override
-            public void onClick(View view) {
-                Intent intent = new Intent(MainActivity.this, CadastroActivity.class);
-                startActivity(intent);
-
-            }
-        });
-
-    }
-}
