--- app/src/main/res/drawable/button_sign.xml
@@ -1,9 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape= "rectangle">
-
-    <corners android:radius="8dp" />
-
-    <solid android:color="#D12876CA" />
-
-</shape>
