+++ app/src/main/res/drawable/ic_google_logo.xml
@@ -1,0 +1,10 @@
+<vector xmlns:android="http://schemas.android.com/apk/res/android"
+    android:width="24dp"
+    android:height="24dp"
+    android:viewportWidth="512"
+    android:viewportHeight="512">
+    <path
+        android:fillColor="#FFFFFF"
+        android:pathData="m260.743,222.339l221.86,0q3.672,20.503 3.672,39.17q0,66.405 -27.847,118.58t-79.41,81.553t-118.274,29.377q-48.044,0 -91.498,-18.514t-74.973,-50.033t-50.033,-74.973t-18.514,-91.498t18.514,-91.498t50.033,-74.973t74.973,-50.033t91.498,-18.514q91.804,0 157.597,61.509l-63.957,61.509q-37.64,-36.416 -93.64,-36.416q-39.476,0 -72.984,19.891t-53.093,54.011t-19.585,74.514t19.585,74.514t53.093,54.011t72.984,19.891q26.623,0 48.962,-7.344t36.722,-18.361t25.093,-25.093t15.76,-26.623t6.885,-23.869l-133.422,0l0,-80.788z"
+        android:strokeWidth="1" />
+</vector>
