+++ app/src/main/res/drawable/shape_google_button.xml
@@ -2,8 +2,7 @@
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
     android:shape= "rectangle">
 
-    <corners android:radius="8dp" />
+    <corners android:radius="@dimen/radius_10" />
 
-    <solid android:color="#D11948E1" />
-
+    <solid android:color="@color/google_color" />
 </shape>
