+++ app/src/main/res/drawable/shape_normal_button.xml
@@ -1,8 +1,8 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android"
-    android:shape= "rectangle">
+    android:shape="rectangle">
 
     <corners android:radius="@dimen/radius_10" />
 
-    <solid android:color="#1C2732" />
+    <solid android:color="#2876CA" />
 </shape>
