+++ app/src/main/res/layout/activity_splash.xml
@@ -11,29 +11,28 @@
         android:id="@+id/textView"
         android:layout_width="wrap_content"
         android:layout_height="wrap_content"
-        android:layout_marginStart="30dp"
-        android:layout_marginEnd="30dp"
-        android:layout_marginBottom="16dp"
-        android:text="NASA Experience"
+        android:layout_marginStart="@dimen/margin_16"
+        android:layout_marginEnd="@dimen/margin_16"
+        android:layout_marginBottom="@dimen/margin_16"
+        android:text="@string/app_name"
         android:textAlignment="center"
-        android:textColor="@android:color/background_light"
-        android:textSize="38sp"
+        android:textColor="@color/white"
+        android:textSize="@dimen/font_32"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.5"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toBottomOf="@+id/imageView4" />
 
     <ImageView
         android:id="@+id/imageView4"
-        android:layout_width="188dp"
-        android:layout_height="170dp"
-        android:layout_marginTop="50dp"
-        android:scaleType="centerInside"
+        android:layout_width="@dimen/width_200"
+        android:layout_height="@dimen/height_200"
+        android:layout_marginTop="@dimen/margin_16"
+        android:importantForAccessibility="no"
+        android:scaleType="fitCenter"
         app:layout_constraintBottom_toTopOf="@+id/textView"
         app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintHorizontal_bias="0.499"
         app:layout_constraintStart_toStartOf="parent"
         app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/ic_launcher_nasa" />
+        app:srcCompat="@drawable/ic_nasa_logo" />
 </androidx.constraintlayout.widget.ConstraintLayout>
